Change the way how document change is detected
Before this change PixiEditor detected a change in a document by
trusting events fired from Bitmap and Layer. Trusting to those
events is naive as they can be fired when there is no change
and user is then prompted for unsaved changes without reason.

Changed this so that that copy of the original document is saved
and when a user is trying to exit without saving this the active
document is compared with original and change is detected that way.

This should fix issues #50 and #49.